# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is the admin API for the Urchin app.",
        "title": "GoCMS Admin API",
        "contact": {
            "name": "Ricardo",
            "email": "ricardobenthem@gmail.com"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8081",
    "basePath": "/",
    "paths": {
...
        "/pages": {
            "post": {
                "description": "Adds a new page to the database.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pages"
                ],
                "summary": "Add a new page",
                "parameters": [
                    {
                        "description": "Page to Update",
                        "name": "page",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/admin_app.ChangePageRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/admin_app.PageResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body or data",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/posts/{id}": {
            "get": {
                "description": "Retrieves a single post by its ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "posts"
                ],
                "summary": "Get a single post",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Post ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/admin_app.GetPostResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid post ID",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Post not found",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes a page by its Link.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pages"
                ],
                "summary": "Delete a page",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Page Link",
                        "name": "link",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/admin_app.DeletePageRequest"
                        }
                    },
                    "400": {
                        "description": "Invalid link provided",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Page not found",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "admin_app.ChangePageRequest": {
            "type": "object",
            "properties": {
                "content": {
                    "description": "Content of the page\nin: body",
                    "type": "string"
                },
                "id": {
                    "description": "ID of the page\nin: body\nrequired: true",
                    "type": "integer"
                },
                "link": {
                    "description": "Link of the page\nin: body",
                    "type": "string"
                },
                "title": {
                    "description": "Title of the page\nin: body",
                    "type": "string"
                }
            }
        },
...
        "admin_app.DeletePageRequest": {
            "type": "object",
            "properties": {
                "link": {
                    "description": "Link of the page to delete\nin: body\nrequired: true",
                    "type": "string"
                }
            }
        },
...
        "admin_app.GetPostResponse": {
            "type": "object",
            "properties": {
                "content": {
                    "description": "Content of the post",
                    "type": "string"
                },
                "excerpt": {
                    "description": "Excerpt of the post",
                    "type": "string"
                },
                "id": {
                    "description": "ID of the post",
                    "type": "integer"
                },
                "title": {
                    "description": "Title of the post",
                    "type": "string"
                }
            }
        },
...
        "admin_app.PageResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "ID of the page",
                    "type": "integer"
                },
                "link": {
                    "description": "Link of the page",
                    "type": "string"
                }
            }
        },
...
        "common.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
...
    }
}